+++ R/plots.R
@@ -1,9 +1,9 @@
 #' Plot RLFS analysis results
 #'
-#' Plots the results of the R-loop-forming sequences (RLFS) analysis. The plot is a 
-#' metaplot of the Z score distribution around RLFS with the p value from 
+#' Plots the results of the R-loop-forming sequences (RLFS) analysis. The plot is a
+#' metaplot of the Z score distribution around RLFS with the p value from
 #' permutation testing annotated. See also [analyzeRLFS].
-#' 
+#'
 #' @param object An RLRanges object with [analyzeRLFS] already run.
 #' Alternatively, can be the RLFS results object from an RLRanges (from
 #' `rlresult(object, "rlfsRes")`).
@@ -54,7 +54,7 @@ plotRLFSRes <- function(object,
         fft, "Fourier-transformed Z-Score around RLFS", "Z-score around RLFS"
     )
     pltylab <- ifelse(fft, "FFT of Z-score", "Peak Enrichment (Z-Score)")
-    pltxlab <- ifelse(fft, "Frequency", "Distance to RLFS (bp)")
+    pltxlab <- ifelse(fft, "Frequency Domain", "Distance to RLFS (bp)")
     # Make plot
     pltbase <- ggplot2::ggplot(
         data = pltdat,
@@ -71,7 +71,7 @@ plotRLFSRes <- function(object,
             linetype = "dashed"
         ) + ggplot2::scale_y_continuous(expand = c(0, 0))
     }
-    pltbase +
+    pltbase <- pltbase +
         ggplot2::geom_line(size = 1) +
         ggplot2::ggtitle(plttitle) +
         ggplot2::labs(caption = paste0("p < ", round(pval, digits = 5))) +
@@ -79,11 +79,246 @@ plotRLFSRes <- function(object,
         ggplot2::xlab(pltxlab) +
         ggprism::theme_prism(base_size = 14) +
         ggplot2::labs(subtitle = plotName)
+    if (fft) {
+        # Remove xaxis ticks / numbers if plotting fourier frequency
+        pltbase <- pltbase +
+            ggplot2::theme(
+                axis.text.x = ggplot2::element_blank(),
+                axis.ticks.x = ggplot2::element_blank()
+            )
+    }
+    return(pltbase)
+}
+
+
+#' Plot noise analysis results as a fingerprint plot
+#'
+#' Plots the results of the noise analysis in [noiseAnalyze]. Creates a
+#' fingerprint plot like those developed by
+#' [Diaz et al, 2012](https://pubmed.ncbi.nlm.nih.gov/22499706/) and those
+#' provided by
+#' [deepTools](https://deeptools.readthedocs.io/en/develop/content/tools/plotFingerprint.html).
+#'
+#' The term "Fingerprint plot" comes from deepTools.
+#'
+#' @param object An RLRanges object with [noiseAnalyze] already run.
+#' @return A ggplot object. See also [ggplot2::ggplot].
+#'
+#' 2. `noiseComparisonPlot`
+#'   - A plot showing the noise analysis
+#'   results from the user-supplied sample compared to similar samples from
+#'   RLBase.
+#'
+#' @examples
+#'
+#' # Example RLRanges dataset with analyzeRLFS() already run.
+#' rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))
+#'
+#' # Plot RLFS res
+#' plotFingerprint(rlr)
+#'
+#' @export
+plotFingerprint <- function(object) {
+
+    # Get the noise analysis results
+    rlnoise <- rlresult(object, "noiseAnalysis")
+
+    # Get sample name
+    sname <- object@metadata$sampleName
+
+    # Create a fingerprint plot
+    fingerprint <- rlnoise %>%
+        dplyr::mutate(
+            rank = .data$rank / max(.data$rank)
+        ) %>%
+        ggplot2::ggplot(
+            ggplot2::aes_string(
+                x = "rank",
+                y = "value"
+            )
+        ) +
+        ggplot2::geom_abline(
+            slope = 1, intercept = 0,
+            color = "grey",
+            linetype = "dashed",
+            alpha = .5
+        ) +
+        ggplot2::geom_line() +
+        ggplot2::labs(
+            title = paste0("Fingerprint plot of ", sname)
+        ) +
+        ggplot2::xlab("Proportion of bins") +
+        ggplot2::ylab("Proportion of signal") +
+        ggplot2::theme_linedraw(
+            base_size = 14
+        ) +
+        ggplot2::theme(
+            plot.caption = ggplot2::element_text(size = 8)
+        )
+
+    return(fingerprint)
+}
+
+
+#' Creates a metaplot for comparing noise analysis results with RLBase
+#'
+#' Plots the average standardized signal from [noiseAnalyze] alongside the
+#' samples in RLBase. For this plot, lower average signal indicates better
+#' signal to noise ratio. **Note**: This plot may be misleading if you supplied
+#' custom windows when running [noiseAnalyze].
+#'
+#' @param object An RLRanges object with [noiseAnalyze] already run.
+#' @param mode A `character` containing the R-loop data mode to compare
+#' against. See details for more information.
+#' @param simple A `logical` which specifies whether the plot should only show
+#' samples where the prediction and label are the same. Default: TRUE.
+#' @param returnData If TRUE, plot data is returned instead of plotting.
+#'  Default: FALSE
+#' @return A [ggplot2::ggplot] object or a `tbl` if `returnData` is `TRUE`.
+#' @details
+#'
+#' ## Mode
+#'
+#' The `mode` parameter specifies the R-loop modality to compare the
+#' user-supplied sample against in the plot. The default, `"auto"`
+#' specifies that the `mode` from the supplied `RLRanges` object will
+#' be used. Only one mode can be specified. For a list of applicable modes,
+#' see `auxdata$available_modes`.
+#'
+#' ## Plot
+#'
+#' The plot is a violin / jitter plot showing the distribution of average values
+#' from the [noiseAnalyze] output across RLBase samples of the selected `mode`.
+#' The user-supplied sample is annotated on the plot.
+#'
+#' @examples
+#'
+#' rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))
+#'
+#' # Plot RL-Region overlap
+#' noiseComparisonPlot(rlr)
+#'
+#' # Return data only
+#' noiseComparisonPlot(rlr, returnData = TRUE)
+#' @export
+noiseComparisonPlot <- function(object, mode = "auto", simple = TRUE, returnData = FALSE) {
+
+    # Retrieve results
+    noiseres <- rlresult(object, "noiseAnalysis")
+
+    # Get the noise index
+    noiseindex <- mean(noiseres$value)
+
+    # Wrangle along with sample info
+    userdata <- dplyr::tibble(
+        sample = object@metadata$sampleName,
+        noise_index = noiseindex,
+        label = object@metadata$label,
+        prediction = object@metadata$results@predictRes$prediction,
+        group = "User-supplied"
+    )
+
+    # Get genome
+    genome <- GenomeInfoDb::genome(object)[1]
+
+    # Get rlbase samps
+    rlsamples <- RLHub::rlbase_samples()
+
+    # Get the mode
+    if (mode == "auto") mode <- object@metadata$mode
+
+    # Get the RLBase noise analysis results and wrangle them
+    toplt <- rlbaseNoiseAnalyze %>%
+        dplyr::inner_join(rlsamples, by = "rlsample") %>%
+        dplyr::filter(
+            .data$mode == {{ mode }},
+            .data$genome == {{ genome }}
+        ) %>%
+        dplyr::mutate(group = {{ mode }}) %>%
+        dplyr::select(
+            sample = .data$rlsample,
+            noise_index = .data$value,
+            .data$label, .data$prediction, .data$group
+        ) %>%
+        dplyr::bind_rows(userdata)
+
+    if (returnData) {
+        return(toplt)
+    }
+
+    # Plot
+    if (!simple) {
+        cond <- "cond"
+        colvec <- auxdata$prediction_label_cols
+        xlab <- "Sample QC prediction/label (prediction_label)"
+        title <- "Noise index across prediction-label combinations"
+        topltfinal <- toplt
+    } else {
+        cond <- "prediction"
+        colvec <- stats::setNames(
+            auxdata$prediction_cols$col,
+            nm = auxdata$prediction_cols$prediction
+        )
+        xlab <- "Sample QC prediction"
+        topltfinal <- toplt %>%
+            dplyr::filter(
+                .data$prediction == .data$label | .data$group == "User-supplied"
+            )
+        title <- "Noise comparison plot"
+    }
+    plt <- topltfinal %>%
+        dplyr::mutate(
+            cond = paste0(.data$prediction, "_", .data$label),
+            cond = factor(.data$cond, levels = c("NEG_NEG", "NEG_POS", "POS_NEG", "POS_POS"))
+        ) %>%
+        ggplot2::ggplot(ggplot2::aes_string(x = cond, fill = cond, y = "noise_index")) +
+        ggplot2::geom_boxplot(
+            width = ifelse(simple, .8, 1),
+            position = ggplot2::position_dodge(width = .9),
+            outlier.shape = NA
+        ) +
+        ggplot2::geom_point(
+            position = ggplot2::position_jitterdodge(
+                dodge.width = .9,
+                jitter.width = ifelse(simple, 0.25, 1),
+                seed = 42
+            )
+        ) +
+        ggplot2::geom_point(
+            alpha = ifelse(topltfinal$group == "User-supplied", 1, 0),
+            size = 4,
+            color = "black",
+            position = ggplot2::position_jitterdodge(
+                dodge.width = .9,
+                jitter.width = 0,
+                jitter.height = 0
+            ),
+            shape = 23,
+            stroke = 1.5
+        ) +
+        ggplot2::scale_y_log10(expand = ggplot2::expansion(mult = c(0, .2))) +
+        ggplot2::theme_bw(base_size = 14) +
+        ggplot2::scale_fill_manual(
+            values = colvec
+        ) +
+        ggplot2::ylab("Noise index (log scale)") +
+        ggplot2::xlab(xlab) +
+        ggplot2::labs(
+            title = title,
+            subtitle = paste0("Data modality: ", mode, "-seq"),
+            caption = paste0(
+                "\u25C7 - User-supplied sample"
+            )
+        ) +
+        ggplot2::theme(legend.position = "none")
+
+    # Return plt
+    return(plt)
 }
 
 
 #' Plot Correlation Results
-#' 
+#'
 #' Plots a heatmap to visualize the pairwise Pearson correlation matrix
 #' generated via [corrAnalyze].
 #'
@@ -132,7 +367,8 @@ corrHeatmap <- function(object,
     annoCorr <- rlsamples %>%
         dplyr::mutate(Selected = "RLBase") %>%
         dplyr::select(
-            .data$rlsample, Mode = .data$mode,
+            .data$rlsample,
+            Mode = .data$mode,
             Label = .data$label,
             Prediction = .data$prediction, .data$Selected
         ) %>%
@@ -252,26 +488,26 @@ corrHeatmap <- function(object,
 #'
 #' Creates a list of plots, one for each annotation database
 #' (see [RLHub::annotations]).
-#' These plots show the feature enrichment for the user-supplied sample in 
-#' comparison to the samples in 
+#' These plots show the feature enrichment for the user-supplied sample in
+#' comparison to the samples in
 #' [RLBase](https://gccri.bishop-lab.uthscsa.edu/rlbase/). This
 #' will only work if you did not use custom annotations with [featureEnrich].
-#' 
+#'
 #' @param object An RLRanges object with [featureEnrich] already run.
-#' @param pred_POS_only If TRUE, only "POS" predicted samples included (see 
-#' also [predictCondition]). Default: TRUE. 
+#' @param pred_POS_only If TRUE, only "POS" predicted samples included (see
+#' also [predictCondition]). Default: TRUE.
 #' @param label_POS_only If TRUE, only "POS" labeled samples included (samples
 #' which are expected to robustly map R-loops,
 #' e.g., "D210N" condition R-ChIP data). Default: FALSE.
 #' @param splitby Metadata by which to split plots. Can be "none", "prediction",
 #'  or "label".
-#' @param limits Specify limits on data range. This is used for controlling 
-#' the infinite estimation of odds ratio resulting from fisher's exact test. 
+#' @param limits Specify limits on data range. This is used for controlling
+#' the infinite estimation of odds ratio resulting from fisher's exact test.
 #' To remove limits, set c(-Inf, Inf). Default: c(-10, 15).
 #' @param returnData If TRUE, plot data is returned instead of plot objects.
 #'  Default: FALSE
 #' @param ... For internal use.
-#' @return A named list of [ggplot2::ggplot] objects. Names correspond to 
+#' @return A named list of [ggplot2::ggplot] objects. Names correspond to
 #' the annotations provided. See also [featureEnrich].
 #' @examples
 #'
@@ -375,7 +611,8 @@ plotEnrichment <- function(object,
     # Make selected explicit
     input_data$selected <- ifelse(input_data$experiment == usamp, usamp, "")
     input_data <- dplyr::distinct(
-        input_data, .data$db, .data$type, .data$experiment, .keep_all = TRUE
+        input_data, .data$db, .data$type, .data$experiment,
+        .keep_all = TRUE
     )
 
     # Build plots
@@ -562,6 +799,125 @@ feature_ggplot <- function(x, usamp, limits, splitby) {
 }
 
 
+#' Plot Transcript Feature Overlap
+#'
+#' Plots the results of [txFeatureOverlap] alongside the average from
+#' public R-loop datasets. This allows comparison of user-supplied
+#' samples with data that is expected to be simialr.
+#'
+#' @param object An RLRanges object with [txFeatureOverlap] already run.
+#' @param mode A `character` containing the R-loop data mode to compare
+#' against. See details for more information.
+#' @param returnData If TRUE, plot data is returned instead of plotting.
+#'  Default: FALSE
+#' @return A [ggplot2::ggplot] object or a `tbl` if `returnData` is `TRUE`.
+#' @details
+#'
+#' ## Mode
+#'
+#' The `mode` parameter specifies the R-loop modality to compare the
+#' user-supplied sample against in the plot. The default, `"auto"`
+#' specifies that the `mode` from the supplied `RLRanges` object will
+#' be used. Only one mode can be specified. For a list of applicable modes,
+#' see `auxdata$available_modes`.
+#'
+#' ## Plot
+#'
+#' The plot is a stacked bar chart showing the proportion of peaks overlapping
+#' transcript features for the supplied `RLRanges` object. Additionally, the
+#' average of the [txFeatureOverlap] analysis for all samples within the
+#' specified modes are also shown as a background comparison.
+#'
+#' This style of analysis enables a user to see the transcript features
+#' overlapping their peaks and compare those results to the average within
+#' relevant public datasets.
+#'
+#' @examples
+#'
+#' rlr <- readRDS(system.file("extdata", "rlrsmall.rds", package = "RLSeq"))
+#'
+#' # Plot RL-Region overlap
+#' plotTxFeatureOverlap(rlr)
+#'
+#' # Return data only
+#' plotRLRegionOverlap(rlr, returnData = TRUE)
+#' @export
+plotTxFeatureOverlap <- function(object, mode = "auto", returnData = FALSE) {
+
+    # Retrieve results
+    txfeatres <- rlresult(object, "txFeatureOverlap")
+
+    # Get genome
+    genome <- GenomeInfoDb::genome(object)[1]
+
+    # Get rlbase samps
+    rlsamples <- RLHub::rlbase_samples()
+
+    # Get the mode
+    if (mode == "auto") mode <- object@metadata$mode
+
+    # Get the rlbase tx feat overlap
+    rlspos <- rlsampleTxOl %>%
+        dplyr::inner_join(rlsamples, by = "rlsample") %>%
+        dplyr::filter(
+            .data$label == "POS",
+            .data$mode == {{ mode }},
+            .data$numPeaks > 5000,
+            .data$prediction == "POS",
+            .data$genome == {{ genome }}
+        ) %>%
+        dplyr::group_by(.data$feature) %>%
+        dplyr::summarise(
+            pct = mean(.data$pct),
+        ) %>%
+        dplyr::mutate(group = paste0({{ mode }}, "-avg"))
+
+    # Get stats for the user-supplied sample
+    olstat <- txfeatres %>%
+        dplyr::group_by(.data$feature) %>%
+        dplyr::tally() %>%
+        dplyr::mutate(
+            pct = .data$n / sum(.data$n),
+            group = object@metadata$sampleName
+        ) %>%
+        dplyr::select(-.data$n)
+    toplt <- olstat %>%
+        dplyr::bind_rows(rlspos) %>%
+        dplyr::mutate(
+            feature = factor(.data$feature, levels = rev(c(
+                "TSS", "fiveUTR", "Exon", "Intron",
+                "threeUTR", "TTS", "Intergenic"
+            )))
+        )
+    if (returnData) {
+        return(toplt)
+    }
+
+    # Plot
+    plt <- toplt %>%
+        ggplot2::ggplot(ggplot2::aes_string(x = "group", y = "pct", fill = "feature")) +
+        ggplot2::geom_col(color = "black") +
+        ggplot2::coord_flip() +
+        ggplot2::theme_bw(base_size = 14) +
+        ggplot2::theme(
+            axis.title.y = ggplot2::element_blank()
+        ) +
+        ggplot2::ylab("Proportion of peaks") +
+        ggplot2::ggtitle(
+            "Transcript Feature Overlap",
+            subtitle = paste0(
+                object@metadata$sampleName,
+                " (user-supplied data) vs RLBase ",
+                paste0(mode, "-avg")
+            )
+        ) +
+        ggplot2::scale_fill_discrete(guide = ggplot2::guide_legend(reverse = TRUE))
+
+    # Return plt
+    return(plt)
+}
+
+
 #' Plot RL-Region overlap with RLRanges
 #'
 #' Convenience function for plotting the overlap between RLRanges and R-loop
@@ -574,7 +930,7 @@ feature_ggplot <- function(x, usamp, limits, splitby) {
 #' Obtained from RLHUb using [RLHub::rlregions_meta]. Loaded from RLHub if
 #' not supplied. Default: NULL.
 #' @param ... Additional arguments passed to [VennDiagram::venn.diagram]
-#' @return A [ggplot2::ggplot] object containing the venn diagram. Built 
+#' @return A [ggplot2::ggplot] object containing the venn diagram. Built
 #' using [ggplotify::as.ggplot].
 #' @examples
 #'
@@ -646,10 +1002,11 @@ plotRLRegionOverlap <- function(object, returnData = FALSE,
     # Suppress output log from venn.diagram if futile logger available
     if (requireNamespace("futile.logger", quietly = TRUE)) {
         futile.logger::flog.threshold(
-            futile.logger::ERROR, name = "VennDiagramLogger"
+            futile.logger::ERROR,
+            name = "VennDiagramLogger"
         )
     }
-    
+
     # Make the plot
     gt <- pltdata %>%
         VennDiagram::venn.diagram(